+++ packages/expo-router/CHANGELOG.md
@@ -22,6 +22,8 @@
 - fix screen freeze after preview navigation ([#37881](https://github.com/expo/expo/pull/37881) by [@Ubax](https://github.com/Ubax))
 - fix detents in Modal ([#37981](https://github.com/expo/expo/pull/37981) by [@Ubax](https://github.com/Ubax))
 - fix(web): a11y focus trap issues in web modals ([#38026](https://github.com/expo/expo/pull/38026) by [@hirbod](https://github.com/hirbod))
+- Fix web modal styling ([#38040](https://github.com/expo/expo/pull/38040) by [@Ubax](https://github.com/Ubax))
+
 
 ### 💡 Others